pyproject.toml
[project]
name = "langchain-tour"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.0.5",
    "langchain-deepseek>=1.0.1",
    "langchain-mistralai>=1.0.1",
    "langchain-openai>=1.0.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
